Split out the command expansion tests into their own file
